+++ common/gl_rsurf.c
@@ -438,7 +438,7 @@ DrawFlatGLPoly(glpoly_t *p)
     int i;
     float *v;
 
-    srand((unsigned long)p);
+    srand((intptr_t)p);
     glColor3f(rand() % 256 / 255.0, rand() % 256 / 255.0,
 	      rand() % 256 / 255.0);
     glBegin(GL_POLYGON);
